Use multiband_image for remote sensing group
