+++ api/model/storage/postgres/bounds.go
@@ -371,6 +371,10 @@ func (f *BoundsField) parseHistogram(rows pgx.Rows, xExtrema *api.Extrema, yExtr
 		}
 		xBuckets[xBucket].Buckets[yBucket].Count += yRowBucketCount
 	}
+	err := rows.Err()
+	if err != nil {
+		return nil, errors.Wrapf(err, "error reading data from postgres")
+	}
 
 	// assign histogram attributes
 	return &api.Histogram{
