+++ api/model/storage/postgres/variable.go
@@ -203,7 +203,7 @@ func (s *Storage) fetchSummaryData(dataset string, storageName string, varName s
 			field = NewCoordinateField(variable.Name, s, dataset, storageName, gcg.XCol, gcg.YCol, variable.DisplayName, variable.Grouping.GetType(), "")
 		} else if model.IsRemoteSensing(variable.Grouping.GetType()) {
 			rsg := variable.Grouping.(*model.RemoteSensingGrouping)
-			field = NewRemoteSensingField(variable.Name, s, dataset, storageName, rsg.CoordinateCol, variable.DisplayName, variable.Grouping.GetType(), model.D3MIndexName)
+			field = NewMultiBandImageField(s, dataset, storageName, variable.Name, variable.DisplayName, variable.Grouping.GetType(), rsg.IDCol, rsg.BandCol)
 		} else {
 			return nil, errors.Errorf("variable grouping `%s` of type `%s` does not support summary", variable.Grouping.GetIDCol(), variable.Grouping.GetType())
 		}
@@ -234,12 +234,10 @@ func (s *Storage) fetchSummaryData(dataset string, storageName string, varName s
 			field = NewTextField(s, dataset, storageName, variable.Name, variable.DisplayName, variable.Type, countCol)
 		} else if model.IsImage(variable.Type) {
 			field = NewImageField(s, dataset, storageName, variable.Name, variable.DisplayName, variable.Type, countCol)
-		} else if model.IsMultiBandImage(variable.Type) {
-			field = NewMultiBandImageField(s, dataset, storageName, variable.Name, variable.DisplayName, variable.Type, countCol)
 		} else if model.IsDateTime(variable.Type) {
 			field = NewDateTimeField(s, dataset, storageName, variable.Name, variable.DisplayName, variable.Type, countCol)
 		} else if model.IsGeoBounds(variable.Type) {
-			field = NewGeoBoundsField(s, dataset, storageName, variable.Name, variable.DisplayName, variable.Type, countCol)
+			field = NewBoundsField(s, dataset, storageName, variable.Name, variable.DisplayName, variable.Type, countCol)
 		} else {
 			return nil, errors.Errorf("variable `%s` of type `%s` does not support summary", variable.Name, variable.Type)
 		}
